v3.6.19
Add comprehensive tests for various features and services

- Implement regression tests for bootstrap AppState wiring in `useAppBootstrap`.
- Create tests for `plansService` to validate API interactions for active and completed strategies.
- Add extensive tests for the `useHealthData` hook, covering platform dispatching and date range syncing.
- Introduce tests for `ApiClient` to ensure proper error handling and normalization of API responses.
- Develop tests for OAuth nonce/state generation utilities, validating nonce uniqueness and state management.

+++ packages/contracts/stripe/payment.ts
@@ -126,7 +126,7 @@ export type PaymentHistoryStatus = (typeof PAYMENT_HISTORY_STATUSES)[number];
 export const PaymentHistoryItemSchema = z.object({
   id: z.string(),
   invoiceId: z.string(),
-  amount: z.number(),
+  amount: z.number().int(), // cents
   status: z.enum(PAYMENT_HISTORY_STATUSES),
   date: z.string(),
   invoicePdfUrl: z.string().nullable(),
